refactor(shipping): Phase 5 relocate Carriers + Observability + ICarrierGateway abstraction
- Move 8 Carriers/* files to Infrastructure/Carriers/ and
  Observability/ShippingMetrics.cs to Infrastructure/Observability/.
- Extract ICarrierGateway (+ companion records: ShipmentQuoteRequest,
  CarrierQuote, ShipmentDispatchRequest, CarrierDispatchResult,
  CarrierStatus, CarrierStatusCode, CarrierWebhookUpdate) to
  Domain/Abstractions/ICarrierGateway.cs; implementations stay in
  Infrastructure/Carriers/.
- Rebase 15 consumer usings (Program.cs, ShippingApiEndpoints.cs,
  6 Carriers/* siblings, 7 Shipping.Tests files); build green,
  107/107 Shipping.Tests pass (7 Phase-10 skips unchanged).
