+++ shipping-microservice/Shipping.Service/Infrastructure/Carriers/FakeExpressCarrierGateway.cs
@@ -1,7 +1,8 @@
 using System.Text.Json;
 using Shipping.Service.Domain;
+using Shipping.Service.Domain.Abstractions;
 
-namespace Shipping.Service.Carriers;
+namespace Shipping.Service.Infrastructure.Carriers;
 
 internal sealed class FakeExpressCarrierGateway : ICarrierGateway
 {
