+++ shipping-microservice/Shipping.Service/Infrastructure/Carriers/FakeGroundCarrierGateway.cs
@@ -1,7 +1,8 @@
 using System.Text.Json;
 using Shipping.Service.Domain;
+using Shipping.Service.Domain.Abstractions;
 
-namespace Shipping.Service.Carriers;
+namespace Shipping.Service.Infrastructure.Carriers;
 
 internal sealed class FakeGroundCarrierGateway : ICarrierGateway
 {
