+++ shipping-microservice/Shipping.Service/Program.cs
@@ -7,12 +7,13 @@ using ECommerce.Shared.IntegrationEvents.Commands;
 using ECommerce.Shared.Observability;
 using ECommerce.Shared.OpenApi;
 using ECommerce.Shared.Qa;
-using Shipping.Service.Carriers;
 using Shipping.Service.Contracts.Integration;
+using Shipping.Service.Domain.Abstractions;
 using Shipping.Service.Endpoints;
+using Shipping.Service.Infrastructure.Carriers;
 using Shipping.Service.Infrastructure.Data.EntityFramework;
+using Shipping.Service.Infrastructure.Observability;
 using Shipping.Service.IntegrationEvents.EventHandlers;
-using Shipping.Service.Observability;
 
 var builder = WebApplication.CreateBuilder(args);
