+++ shipping-microservice/Shipping.Tests/Carriers/RateShoppingServiceTests.cs
@@ -1,5 +1,6 @@
-using Shipping.Service.Carriers;
 using Shipping.Service.Domain;
+using Shipping.Service.Domain.Abstractions;
+using Shipping.Service.Infrastructure.Carriers;
 
 namespace Shipping.Tests.Carriers;
