+++ shipping-microservice/Shipping.Tests/IntegrationEvents/EventStreamConsolidationTests.cs
@@ -7,9 +7,10 @@ using ECommerce.Shared.IntegrationEvents.Commands;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.ApiModels;
-using Shipping.Service.Carriers;
 using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Domain;
+using Shipping.Service.Domain.Abstractions;
+using Shipping.Service.Infrastructure.Carriers;
 using Shipping.Service.Infrastructure.Data.EntityFramework;
 
 namespace Shipping.Tests.IntegrationEvents;
