+++ shipping-microservice/Shipping.Tests/Observability/ShippingMetricsTests.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics.Metrics;
 using ECommerce.Shared.Observability.Metrics;
 using Shipping.Service.Domain;
-using Shipping.Service.Observability;
+using Shipping.Service.Infrastructure.Observability;
 
 namespace Shipping.Tests.Observability;
